split hqadmin tests into a tests dir
