+++ corehq/apps/hqadmin/tests/__init__.py
@@ -1,0 +1,2 @@
+from test_prbac import *
+from test_pillowtop import *
